Move the test out of test.prettyprinter
