+++ src/test/java/spoon/test/issue6665/testclasses/SampleClassIssue6665.java
@@ -1,4 +1,4 @@
-package spoon.test.prettyprinter.testclasses;
+package spoon.test.issue6665.testclasses;
 
 public class SampleClassIssue6665 {
 	public static void main(String[] args) {}
